tests: mock entrypoints tests
