+++ jsonresolver/contrib/jsonref.py
@@ -16,62 +16,23 @@ by your ``JSONResolver`` instance.
 Example:
 .. code-block:: python
 
-   >>> from jsonref import JsonRef
+   >>> from jsonref import replace_refs
    >>> from jsonresolver import JSONResolver
    >>> from jsonresolver.contrib.jsonref import json_loader_factory
    >>> schema = {'$ref': 'http://localhost:4000/schema/authors.json#'}
    >>> json_resolver = JSONResolver(plugins=['tests.demo.schema'])
    >>> loader_cls = json_loader_factory(json_resolver)
    >>> loader = loader_cls(cache_results=False)
-   >>> dict(JsonRef.replace_refs(schema, loader=loader))
+   >>> dict(replace_refs(schema, loader=loader))
    {'type': 'array'}
 
    # if you do not want default fallback:
-   >>> dict(JsonRef.replace_refs(schema, loader=json_resolver.resolve))
+   >>> dict(replace_refs(schema, loader=json_resolver.resolve))
    {'type': 'array'}
 
 """
-from collections.abc import MutableMapping
-from urllib import parse as urlparse
-
+from jsonref import URIDict
 from werkzeug.exceptions import NotFound
-
-
-class _URIDict(MutableMapping):
-    """Dictionary which uses normalized URIs as keys."""
-
-    def normalize(self, uri):
-        """Normalize."""
-        return urlparse.urlsplit(uri).geturl()
-
-    def __init__(self, *args, **kwargs):
-        """Construct."""
-        self.store = dict()
-        self.store.update(*args, **kwargs)
-
-    def __getitem__(self, uri):
-        """GetItem."""
-        return self.store[self.normalize(uri)]
-
-    def __setitem__(self, uri, value):
-        """SetItem."""
-        self.store[self.normalize(uri)] = value
-
-    def __delitem__(self, uri):
-        """DelItem."""
-        del self.store[self.normalize(uri)]
-
-    def __iter__(self):
-        """Iterate."""
-        return iter(self.store)
-
-    def __len__(self):
-        """Len."""
-        return len(self.store)
-
-    def __repr__(self):
-        """Repr."""
-        return repr(self.store)
 
 
 class _JsonLoader:
@@ -90,7 +51,7 @@ class _JsonLoader:
 
     def __init__(self, store=(), cache_results=True):
         """Construct."""
-        self.store = _URIDict(store)
+        self.store = URIDict(store)
         self.cache_results = cache_results
 
     def __call__(self, uri, **kwargs):
