+++ setup.cfg
@@ -26,7 +26,7 @@
 [metadata]
 name = jsonresolver
 version = attr: jsonresolver.__version__
-description = "JSON data resolver with support for plugins."
+description = JSON data resolver with support for plugins.
 long_description = file: README.rst, CHANGES.rst
 keywords =
 license = BSD
@@ -43,24 +43,25 @@ packages = find:
 python_requires = >=3.7
 zip_safe = False
 install_requires =
-    six>=1.12.0
     pluggy>=0.10.0,<1.0
     werkzeug>=1.0.0
 
 [options.extras_require]
 tests =
-    pytest-black>=0.3.0,<0.3.10
+    pytest-black>=0.3.0
     mock>=1.3.0
     pytest-cache>=1.0
     requests>=2.7.0
-    pytest-invenio>=1.4.0
+    pytest-invenio>=2.0.0,<3.0.0
     sphinx>=4.5
-    jsonref>=0.1
+    jsonref>=1.0.0
     jsonschema>=2.5.1
 # Kept for backwards compatibility
 docs =
 jsonref =
+    jsonref>=1.0.0
 jsonschema =
+    jsonschema>=2.5.1
 
 [build_sphinx]
 source-dir = docs/
@@ -93,6 +94,8 @@ profile=black
 [check-manifest]
 ignore =
     *-requirements.txt
+    tests/build/lib/demo/*
+    tests/dist/*
 
 [tool:pytest]
 addopts = --black --isort --pydocstyle --doctest-glob="*.rst" --doctest-modules --cov=jsonresolver --cov-report=term-missing
