+++ tests/test_jsonref.py
@@ -10,7 +10,7 @@
 """Define unit tests to test core functionality."""
 
 import pytest
-from jsonref import JsonRef, JsonRefError
+from jsonref import JsonRefError, replace_refs
 
 from jsonresolver import JSONResolver
 from jsonresolver.contrib.jsonref import json_loader_factory
@@ -30,7 +30,7 @@ def test_key_resolver():
     json_resolver = JSONResolver(plugins=["demo.schema"])
     loader_cls = json_loader_factory(json_resolver)
     loader = loader_cls()
-    data = JsonRef.replace_refs(example_schema, loader=loader)
+    data = replace_refs(example_schema, loader=loader)
     assert data["properties"]["authors"] == {"type": "array"}
 
 
@@ -46,13 +46,13 @@ def test_missing_route():
     json_resolver = JSONResolver(plugins=["demo.schema"])
     loader_cls = json_loader_factory(json_resolver)
     loader = loader_cls()
-    data = JsonRef.replace_refs(example_schema, loader=loader)
+    data = replace_refs(example_schema, loader=loader)
     with pytest.raises(JsonRefError):
         data["properties"]["authors"]["type"]
 
 
 def test_same_route_different_hosts():
-    """Test orignal resolver."""
+    """Test original resolver."""
     example = {
         "host1": {"$ref": "http://localhost:4000/test"},
         "host2": {"$ref": "http://inveniosoftware.org/test"},
@@ -61,6 +61,6 @@ def test_same_route_different_hosts():
     json_resolver = JSONResolver(plugins=["demo.simple", "demo.hosts"])
     loader_cls = json_loader_factory(json_resolver)
     loader = loader_cls()
-    data = JsonRef.replace_refs(example, loader=loader)
+    data = replace_refs(example, loader=loader)
     assert data["host1"]["test"] == "test"
     assert data["host2"]["test"] == "inveniosoftware.org"
